requirements.txt
asttokens==2.4.1
connected-components-3d==3.18.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
distlib==0.3.7
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.12.2
fonttools==4.53.1
fsspec==2024.6.1
gudhi==3.10.1
h5py==3.11.0
imageio==2.35.1
ipython==8.26.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy-loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
MedPy==0.5.2
monai==1.3.2
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==3.9.1
POT==0.9.4
prompt-toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
SimpleITK==2.3.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.8.24
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
traitlets==5.14.3
triton==2.3.1
typing-extensions==4.12.2
tzdata==2024.1
UNKNOWN==0.0.0
virtualenv==20.24.1
wcwidth==0.2.13
